# docs-site/openapi.yaml
openapi: 3.0.3
info:
  title: LogPulse API
  version: 1.1.0
  description: |
    REST API for LogPulse — a centralized log monitoring platform.

    Three authentication realms exist:

    | Realm | Mechanism | Used by |
    |-------|-----------|---------|
    | Dashboard JWT | `Authorization: Bearer <jwt>` (24h expiry, HS256) | everything under `/api/**` except ingest |
    | Ingest API key | `Authorization: Bearer <application API key>` | `POST /api/logs` only |
    | OAuth redirect | browser 302 flow | `/api/auth/{google,github}/*` |

    There is no public signup: an admin invites users, and the invited person
    activates the account via `POST /api/auth/accept-invite` or by signing in
    with Google/GitHub.

    Errors are always `{"error": "<human readable message>"}`.
servers:
  - url: http://localhost:8080
    description: Local development
  # INFERRED: production URL depends on your deployment; set your own.
  - url: https://your-logpulse-host
    description: Production (replace with your host)
...
paths:
...
  /api/auth/accept-invite:
    post:
      tags: [Auth]
      summary: Activate an invited account
      description: >
        Consumes the one-time invite token (valid 7 days), sets name and
        password, flips the account to active and returns a JWT.
      operationId: acceptInvite
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, name, password]
              properties:
                token: { type: string, description: Raw token from the invite link }
                name: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Account activated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400':
          description: Invalid/used/expired invite link, or validation error
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard JWT from /api/auth/login — 24h expiry (HS256).
    ingestApiKey:
      type: http
      scheme: bearer
      description: Per-application API key from POST /api/applications.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [admin, editor, viewer] }
        status: { type: string, enum: [invited, active, deactivated] }
        provider: { type: string, enum: [local, google, github] }
        avatar_url: { type: string }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
      # Password hash, provider id, and invite token fields are never returned
      # (userResponse struct in internal/handlers/auth_handler.go).
    LoginResponse:
      type: object
      properties:
        token: { type: string, description: JWT, expires in 24h }
        user: { $ref: '#/components/schemas/User' }
...